# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/shopper_parties:
    get:
      tags:
        - ShopperParties
      summary: List ShopperParties
      description: Get a paginated list of ShopperParties with optional filtering and sorting
      operationId: listShopperParties
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopperPartyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ShopperParties
      summary: Create ShopperParty
      description: Create a new ShopperParty
      operationId: createShopperParty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShopperPartyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopperParty'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/storefront_audit_logs:
    get:
      tags:
        - StorefrontAuditLogs
      summary: List StorefrontAuditLogs
      description: Get a paginated list of StorefrontAuditLogs with optional filtering and sorting
      operationId: listStorefrontAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorefrontAuditLogList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - StorefrontAuditLogs
      summary: Create StorefrontAuditLog
      description: Create a new StorefrontAuditLog
      operationId: createStorefrontAuditLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStorefrontAuditLogRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorefrontAuditLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ShopperParty:
      type: object
      required:
        - id
        - company_id
        - email_normalized
        - party_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        email_normalized:
          type: string
          maxLength: 320
        party_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateShopperPartyRequest:
      type: object
      required:
        - company_id
        - email_normalized
        - party_id
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        email_normalized:
          type: string
          maxLength: 320
        party_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    ShopperPartyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShopperParty'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    StorefrontAuditLog:
      type: object
      required:
        - id
        - event
        - occurred_at
      properties:
        id:
          type: string
          format: uuid
        website_id:
          nullable: true
          type: string
          format: uuid
        event:
          type: string
        actor:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
          maxLength: 120
        subject_id:
          nullable: true
          type: string
          format: uuid
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
...
    CreateStorefrontAuditLogRequest:
      type: object
      required:
        - event
        - occurred_at
      properties:
        website_id:
          nullable: true
          type: string
          format: uuid
        event:
          type: string
        actor:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
          maxLength: 120
        subject_id:
          nullable: true
          type: string
          format: uuid
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
...
    StorefrontAuditLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StorefrontAuditLog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []